requirements.txt
Flask==2.0.3
torch==2.0.0
torchvision==0.15.0
tensorflow==2.10.0
Pillow==9.2.0
numpy==1.21.6
